pkg/operator: Deflate the update logic in handleStorageLocation into its own method
Update pkg/operator/storagelocations.go, deflating the updating logic that is stored in the op.handleStorageLocation method, into it's own method.

This way, we can handle the validation and update logic for any particular StorageLocation the reporting-operator is currently processing into a dedicated method, and reduce the responsibilities of the op.handleStorageLocation driver method to making any of the necessary API calls to the Kubernetes client.