Dockerfile.src: Avoid using the latest Helm image tag
Update the Dockerfile.src (build root Dockerfile for CI tests) and use a release-versioned Helm image tag instead of the `latest` tag.

This is needed as the Helm version has recently been bumped to use the 3.x versioning, which comes with some slight breaking changes to the templating logic.

Note: this commit needs to be force-merged in order to have these changes propagate to CI tests.