charts/openshift-metering: Update the hive-scripts entrypoint script to run the schematool on non-default databases
Update the hive-scripts ConfigMap entrypoint.sh script, which gets
called in the Hive metastore and server StatefulSets and is responsible
for starting those containers.

In the case of derby (the default) databases, using this tool isn't
necessary and will cause issues when running it against an already
exists schema. This tool is needed in order to get metering working
again with MySQL and PostgreSQL versions.