images: Override the default derby database type when postgres or mysql has been configured
Update the Ansible role's `configure_hive_metastore.yaml` task file to
override the `_meteringconfig_hive_metastore_db_url` helper variable.
That helper variable, which is set to the `hive.spec.config.db.url`
configuration and will be overriden with any value that we set in that
task file before the MeteringConfig values have been finalized.

When we're handling postgres resources, we need to remap what we pull
from the JDBC URL as `postgresql` isn't a valid `schematool -dbType`
option, but `postgres` is so use the value from the split URL array
unless we're configuring a postgres database for Hive metastore.