charts/openshift-metering: Add more consistency to the PVC-related queries
Update charts/openshift-metering and attempt to make the SQL queries
detailed in the `ReportQuery` YAML manifests more consistent:
- Add the missing `type: time` inputs to the Reporting* columns.
- Noramlize the tabbing for the columns SELECT-ed for given queries that
  were mis-aligned.
- Add missing column names for columns used in aggregate functions.

For the last bullet point, the `namespace-persistentvolumeclaim-usage`
ReportQuery was using `sum(persistentvolumeclaim_usage_bytes)` but no
alias was used, which resulted in column names, like `_col3`. This is
mainly a no-op when creating reports referencing this `ReportQuery`, but
noticable when running this query against a Presto instance locally.