test/e2e: Fix the report expiration subtest handler functions
Update these report expiration subtest functions that are responsible
for testing the expiration functionality. Before, some of these report
objects that were created were flat out invalid, especially in the case
of the testing the rollup report scenarios.

In the case of the testing rollup reports, we need to create a
reportquery custom resource that acts as the aggregate query and sits
between the sub-report and the aggregate report. Then, we can list that
sub-report as an input to the aggregate report while still referencing a
valid reportquery custom resource.