pkg/operator: Avoid requeueing an expired Report
Update the Report handler and avoid requeueing a Report that is
reporting an `expired `status. Before, this would result in the
reporting-operator hot looping as this Report custom resource would
continue to get queued up. Instead if we're encountering a Report that
has reached it's expiration date, then we should avoid processing it in
the Report handler function and let the call site determine how to
process that Report further in the next cycle.