test/e2e: Ensure all e2e test namespaces use correct label
Update the MySQL and NFS pre-installation functions to use the correct
namespace label in the form of `name: <namespace_prefix>-metering-testin-ns`.
During the teardown function of the hack/e2e.sh bash script, which
serves as the entrypoint to the e2e suite, we search for any namespaces
that match that label, and attempt to delete them unless the e2e suite
has been run in developer mode.

Previously, the namespaces that were created as a result of these MySQL
and NFS pre-installation functions create a `name` label containing the
actual namespace, and not the prefix supplied to the e2e suite, in the
context of an individual install. This led to those namespaces not being
properly purged at the cleanup call site.

Other changes in this commit include:

- Aligning the grouping of package imports with the rest of the e2e
  package(s).
- Updating the other pre-installation functions to use that
  createTestingNamespace helper function.
- Minor cleanup of unnecessary commits, re-organizing functions, etc.