test/e2e: Avoid using a custom CatalogSource as the pre-upgrade install
Update the upgrade test handler and remove the code that creates and
configures a custom CatalogSource custom resource using a 4.5 index
image that contained a BZ fix that wasn't backported. Remove that code
and any comments detailing why that was necessary. Default back to using
the redhat-operators CatalogSource custom resource as the pre-upgrade
install source.