Handle failed InstallPlan during e2e OLM installations
Update the e2e package to deal with InstallPlan failures that we have no
control over. Currently, OLM will mark an InstallPlan as failed when it
encounters any errors, regardless of the context around that error.

In the context of metering, we're seeing more and more e2e flakes
centered around installations not being able to be installed due to
failed InstallPlan resources. These failed installations are typically
due to CRD conflicts being returned from the apiserver and OLM not
handling that error correctly and the result is metering can't be
installed.

We can workaround this by detecting any failed InstallPlan resources
that we know are due to conflicting resource versions and attempt to
re-install metering in the applicable test handlers.